# bug fix to encode_arm64.s: some registers overwritten in memmove call