@AWSjswinney 2020-11-03
@nigeltao The [change to the Go compiler](https://go-review.googlesource.com/c/go/+/243357) has now been merged and will expose this bug in snappy if we don't merge this fix. Can you take a look or refer someone who can?

Thanks!