@bobrik 2021-02-17
Should there be a test for this? The following does the trick, but there's probably a smaller input:

```go
snappy.Encode(nil, []byte{0xa, 0x5c, 0xa, 0x17, 0xa, 0x8, 0x5f, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x5f, 0x5f, 0x12, 0xb, 0x4d, 0x61, 0x74, 0x63, 0x68, 0x53, 0x65, 0x72, 0x69, 0x65, 0x73, 0xa, 0xf, 0xa, 0x2, 0x64, 0x62, 0x12, 0x9, 0x54, 0x65, 0x6e, 0x4d, 0x69, 0x6e, 0x75, 0x74, 0x65, 0xa, 0x13, 0xa, 0xb, 0x54, 0x75, 0x72, 0x62, 0x69, 0x6e, 0x65, 0x54, 0x79, 0x70, 0x65, 0x12, 0x4, 0x56, 0x31, 0x31, 0x32, 0xa, 0x9, 0xa, 0x4, 0x50, 0x61, 0x72, 0x6b, 0x12, 0x1, 0x31, 0x12, 0x10, 0x9, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0xf0, 0x3f, 0x10, 0xb5, 0xaf, 0xdc, 0xf6, 0xfa, 0x2e, 0xa, 0x5c, 0xa, 0x17, 0xa, 0x8, 0x5f, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x5f, 0x5f, 0x12, 0xb, 0x4d, 0x61, 0x74, 0x63, 0x68, 0x53, 0x65, 0x72, 0x69, 0x65, 0x73, 0xa, 0xf, 0xa, 0x2, 0x64, 0x62, 0x12, 0x9, 0x54, 0x65, 0x6e, 0x4d, 0x69, 0x6e, 0x75, 0x74, 0x65, 0xa, 0x13, 0xa, 0xb, 0x54, 0x75, 0x72, 0x62, 0x69, 0x6e, 0x65, 0x54, 0x79, 0x70, 0x65, 0x12, 0x4, 0x56, 0x31, 0x31, 0x32, 0xa, 0x9, 0xa, 0x4, 0x50, 0x61, 0x72, 0x6b, 0x12, 0x1, 0x32, 0x12, 0x10, 0x9, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0xf0, 0x3f, 0x10, 0xb5, 0xaf, 0xdc, 0xf6, 0xfa, 0x2e, 0xa, 0x5c, 0xa, 0x17, 0xa, 0x8, 0x5f, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x5f, 0x5f, 0x12, 0xb, 0x4d, 0x61, 0x74, 0x63, 0x68, 0x53, 0x65, 0x72, 0x69, 0x65, 0x73, 0xa, 0xf, 0xa, 0x2, 0x64, 0x62, 0x12, 0x9, 0x54, 0x65, 0x6e, 0x4d, 0x69, 0x6e, 0x75, 0x74, 0x65, 0xa, 0x13, 0xa, 0xb, 0x54, 0x75, 0x72, 0x62, 0x69, 0x6e, 0x65, 0x54, 0x79, 0x70, 0x65, 0x12, 0x4, 0x56, 0x31, 0x31, 0x32, 0xa, 0x9, 0xa, 0x4, 0x50, 0x61, 0x72, 0x6b, 0x12, 0x1, 0x33, 0x12, 0x10, 0x9, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0xf0, 0x3f, 0x10, 0xb5, 0xaf, 0xdc, 0xf6, 0xfa, 0x2e, 0xa, 0x5c, 0xa, 0x17, 0xa, 0x8, 0x5f, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x5f, 0x5f, 0x12, 0xb, 0x4d, 0x61, 0x74, 0x63, 0x68, 0x53, 0x65, 0x72, 0x69, 0x65, 0x73, 0xa, 0xf, 0xa, 0x2, 0x64, 0x62, 0x12, 0x9, 0x54, 0x65, 0x6e, 0x4d, 0x69, 0x6e, 0x75, 0x74, 0x65, 0xa, 0x13, 0xa, 0xb, 0x54, 0x75, 0x72, 0x62, 0x69, 0x6e, 0x65, 0x54, 0x79, 0x70, 0x65, 0x12, 0x4, 0x56, 0x31, 0x31, 0x32, 0xa, 0x9, 0xa, 0x4, 0x50, 0x61, 0x72, 0x6b, 0x12, 0x1, 0x34, 0x12, 0x10, 0x9, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0xf0, 0x3f, 0x10, 0xb5, 0xaf, 0xdc, 0xf6, 0xfa, 0x2e})
```

It may be also good to create a new tag now that Go 1.16 is out.