@nigeltao 2021-02-25
> Is it possible to get this fix tagged with v.0.0.3?

Done. Sorry for the delay.